Working Veritesting listener
This does not do quite do veritesting yet, because the expressions I have unrolled lines 34-37 of TestPaths.java into are just a+x and b+y. I will fix this soon. But, it replaces a and b with symbolic expressions and jumps to the end of the unrolled code (line 39 of TestPaths.java)