A completed Veritesting example
Constraints are allowed to use one Comparator, but Expressions are allowed none. We need to construct complex Constraints and Expressions which can use more than one comparator. But this is not available for now.

However, IntegerExpression does support a CMP operator which returns -1 if less than, 0 if equal, and 1 if greater than condition is satisfied. Changing the testMe3 veritesting region to perform this operation allows us to perform veritesting on it using the existing SPF infrastructure.

I will be working on adding complex Constraints and Expressions now though.