Fixing a serious bug with putStatic and putField instructions.
The bug was to do with how we fill up FIELD_OUTPUT holes (which
are created for both putStatic and putField instructions, with the
putStatic ones having FieldInfo.localStackSlot set to -1 and putField
ones having Field.localStackSlot set to the local variable's stack slot
from where the object reference will be read).

For FIELD_OUTPUT holes, we would simply set them to be output variables
of a region and write them into the corresponding field in a veritesting
region. But, if a putStatic or putField for the same field was encountered on both sides
of a branch, then this would cause two writes into the same field, which
is wrong.

We now create a intermediate variable hole and set it inside FIELD_OUTPUT
hole's FieldInfo.writeVal. This is used to lookup the variable to be
written into the field. This intermediate variable takes on a value
depending on which side of the branch was taken.