MWW: added support for region construction and also cleaned up and extended the AST somewhat.  My biggest changes involved:   1). Creating "shadow copies" of instructions, so that we can       incrementally transform them using our expression classes.   2). Changing the visitor framework to be a little more powerful       by introducing a parametric return type.
The system now appears to build complex regions correctly, but I am
sure there are still a few remaining bugs.  However, this is (I think)
a pretty good base to build from.

I was hoping to get a traversal done as well to illustrate how it works
but this will have to wait for another day.  Please look at the jkind
map visitors:
https://github.com/agacek/jkind/blob/master/jkind-common/src/jkind/lustre/visitors/ExprMapVisitor.java
https://github.com/agacek/jkind/blob/master/jkind-common/src/jkind/lustre/visitors/AstMapVisitor.java
and iterator visitors:
https://github.com/agacek/jkind/blob/master/jkind-common/src/jkind/lustre/visitors/ExprIterVisitor.java
https://github.com/agacek/jkind/blob/master/jkind-common/src/jkind/lustre/visitors/AstIterVisitor.java

for guidance on how to do this well.