Doing equivalence-checking without JUnit tests because I am having problems in getting changes to these tests to be recompiled.
If I change the number of parameters to TestVeritestingPerf.testHarness in JPF's configuration and the testHarness method, JPF still complains about the number of parameters in the configuration not matching those in the implementation. These JUnit tests are also difficult to set up because of classpath and IntelliJ issues. Hence, I am moving equivalence-checking to a class of its own.