Verifying expected value of the VeritestingListener.veritestingRegionCount metric + adding HigherOrder2 eq-check
Equivalence-checking can find equivalence if we veritest fewer than expected regions. This can cause us to falsely believe that our veritesting implementation is correct when, in fact, little or no veritesting happened. To guard against such false positives, I am adding the ability to specify an expected value for the number of times regions in a test case will be veritested over all execution paths in the equivalence check. The test program will print out this value and SPF will use such output to check if the expected value is equal to the number of instantiated veritesting regions.

Currently, the two higher-order equivalence checks (HigherOrder1, HigherOrder2) are not satisfying my expectation of the number of times veritesting regions in them should be instantiated.