Make veritestRegionExpectedCount a JPF configuration parameter instead of encoding it in the test's Java class
Printing the expected number of times regions in the test case must be instantiated, capturing this output in SPF is hacky and inelegant. Instead, I am asking equivalence-checking testers to specify this parameter in the test's .jpf file as a "veritestRegionExpectedCount = <value>" property. If this is not instantiated, then this comparison of veritestRegionCount against its expected value will be skipped.