WalaVarExpr objects used as the index expression in a array reference should be "unique"-ified too along with the rest of the ArrayRefVarExpr
This required me to change the interface of CloneableVariable.makeUnique so that it returns the "unique"-ified copy of the CloneableVariable instead of "unique"-ifying that CloneableVariable itself. Since FieldRefVarExpr and ArrayRefVarExpr are subclasses of CloneableVariable, their implementations of makeUnique had to be changed too. I also needed the ability to "unique"-ify a WalaVarExpr if an instance of it appears in the index expression of a ArrayRefVarExpr, so I've moved this unique-ification into a public static method in WalaVarExpr.