WalaVarExpr objects used as the index expression in a array reference should be "unique"-ified too along with the rest of the ArrayRefVarExpr
This required me to change the interface of CloneableVariable.makeUnique so that it returns the "unique"-ified copy of the CloneableVariable instead of "unique"-ifying that CloneableVariable itself. Since FieldRefVarExpr and ArrayRefVarExpr are subclasses of CloneableVariable, their implementations of makeUnique had to be changed too. I thought I may need to "unique"-ify a WalaVarExpr if an instance of it appears in the index expression of a ArrayRefVarExpr, so I moved this unique-ification into a public static method in WalaVarExpr. But it later turned out that WalaVarExpr instances are always unique-ified before we get to the unique-ification for fields and arrays. Therefore, ArrayRefVarExpr.makeUnique will assert on WalaVarExpr index being already unique.