Bringing in the translate-nots simplification done by the SimplifyGreenVisitor into the SimplifyRangerExprVisitor (renamed from SimplifyExprVisitor)
The changes in this commit are the result of an attempt at unifying SimplifyRangerExprVisitor and SimplifyGreenVisitor. The only difference was that the SimplifyRangerExprVisitor wasn't implementing the translate-nots simplification. I've moved that simplification into a method that is now being used by both the SimplifyRangerExprVisitor and the SimplifyGreenVisitor. SimplifyGreenVisitor can now be removed unless there are some other Green-level simplifications that we would like to write into SimplifyGreenVisitor.