Skip summarizing infinite loops during static analysis
This is a problem I ran into with Pad.main(). The current static analysis tries to summarize all regions inside ${jpf-symbc}/build/examples wherein it also finds Pad.main(). Pad.main() contains an infinite loop which is a problem for static analysis. I am now using NatLoop for finding all loops in a Wala IR and detecting an infinite loop if a basic block is the start node of a loop but it doesn't have a conditional branch as its last instruction.