More adapter synthesis changes + changing the way we use classpaths
We should translate a colon-separated string of classpaths into a array of URL objects that is used to create a URLClassLoader that is used to load a class. If a class has methods whose signature involves class types not defined in the class path used in the URLClassLoader, then we have to give up on that class. If this happens for the main target class, then we're not going to be able to do any static analysis using the target class as the starting point. Hence, I am changing the way classpaths are handled so that a URLClassLoader has all the classpaths to start with.