Upgrading the local output invariant to also include the fact that every output in the output table *must* have a corresponding gamma expression-assigning statement in teh region summary
The existing local output invariant is that every gamma expression-assigning statement's left-hand side variable should either be a local output or be the one stack output of the region. With this commit, I am making the local output invariant maintain a one-to-one mapping between local outputs and gamma expression assignments modulo the stack output from a gamma expression assignment that need not be among the local outputs of the region.