Fixing region-ending-on-store issue in WBS (Issue #31)
Regions that have a stack output correctly end on a store instruction as per Wala. SSACFG.BasicBlock has two internal Phi instruction lists -- stackSlotPhis and localPhis -- that each contain phi instructions created for output written to the stack and to a local variable respectively. We can use SSACFG.BasicBlock.stackSlotPhis to detect when a region has stack output. Further, we can use SSACFG.BasicBlock.stackSlotPhis to check if the Wala variable that should be output to the stack as per Java Ranger matches the output of the phi instruction in stackSlotPhis (assuming that stackSlotPhis has only one phi instruction). Finally, if Java Ranger found the region to have a local variable output without any stack output and the local outputs table contains an entry for the phi instruction in stackSlotPhis, then Java Ranger is effectively including the store instruction also in the region summary. This situation makes it ok for the region end to move beyond the store instruction. This commit makes this change happen.

This commit is done assuming that SSACFG.BasicBlock.getStackSlotPhis() is a method that can be called from StaticRegion. I have made the corresponding change in Wala [here](https://github.com/vaibhavbsharma/WALA/commit/7cba1711e13834f00f3cb26c8c7679c5be291db8). Therefore, I am checking in the Wala libraries re-compiled with this change.