An improvement on the fix to store-insn-ending regions.
Regions, whose terminus node contains a stackslotphi (a phi instruction that computes an output value to be written to the stack instead of a local variable), should not write the output value of the stackslotphi into a local variable. They should instead rely on ending on a instruction that will use the stack output computed by the region. This happens in two regions in WBS.update(), one that begins with BB4 and ends with BB7, and another that begins with BB22 and ends with BB28.