Reverting a fix previously made in b6ef641fe2cf545eb39534faa1352f353f1951fd that would cause any region containing a throw instruction to not be statically summarized + fixing the crashing the reversion introduces
The previous fix tried to prevent static analysis from summarizing regions like "assert (x != 0 ? count == x + 3 : count == 3);" by throwing a StaticRegionException if we encountered a region containing a throw. But that change caused all regions containing throw instructions to not be summarized. Oops. That fix is reverted in this commit.

We would still like Java Ranger to not crash when statically summarizing regions like "assert (x != 0 ? count == x + 3 : count == 3);". To that effect, this commit makes a change that makes static analysis throw an exception while summarizing a region when it fails to reconstruct the condition inside a if statement. We expect CreateStaticRegions.createComplexIfCondition to always recover a condition and when this doesn't happen, something went wrong and we should abort that region's summarization. This assertion is present in FieldTest3 in case we want to come back and play with it some more.