Static Gamma expressions with same then- and else-expressions can be simplified to a single expression
An Assignment statement that assigns a gamma expression that chooses the same value on both sides can be converted into an assignment statement that assigns that value to its lhs. In such gamma assignments, the lhs often has a wala variable number that is less than the number of wala variable used in the rhs because such regions lie at the beginning of a while loop where the while loop modifies the local variable present in the gamma expression's rhs. Our `computeOutputVars` sees such gamma assignments and, even though the lhs and rhs correspond to the same stack slot, since the lhs' wala variable number is less than the rhs' wala variable number, thinks that the lhs cannot be an output. Only the maximum wala variable number for a given stack slot is thought of as being an output into that stack slot. Therefore, we fail to recognize the lhs of such a gamma expression assignment statement as being the output to a stack slot. The LocalOutputInvariantVisitor instead thinks that this must be a stack output. However, when we try to instantiate such a region, we find that the region does not end on a stack-consuming instruction and therefore abort that region's instantiation.

To prevent all of this confusion, this simplification prevents the LocalOutputInvariantVisitor from populating the gamma expression assignment statement's lhs from becoming a stack output by preventing the gamma expression from being created in the first place. We still think that the rhs in such a gamma expression is an input, and it gets assigned to the gamma expression assignment statement's lhs but the lhs is never used later and gets simplified out of the instantiated region summary. The instantiated region summary has no output to the lhs' corresponding stack slot because the lhs' wala variable number is still less than the rhs' wala variable number. But that is ok because there is no point in reading the value from the rhs' stack slot and writing it back into the same stack slot.

An example of a region where this whole fiasco happens is "replace.amatch([C[CI)I#56". This region should now be instantiated with running replace with 5 symbolic inputs in mode 5.