Deleting redundant SPF cases from the SPFCasesList inside each region
SPFCases weren't getting removed from the region's spfCaseList when more than one spf case was being substituted by a single "multiple" spf case. This was because the SPFCaseStmt wasn't overriding the `boolean equals(Object)` and the `int hashcode()` methods even though it is used with LinkedHashSet. Adding these methods causes the SPFCaseStmt objects to be removed from SpfCasesPass2Visitor.spfCasesSet. This change is useful when computing the number of SPFCase choices in a instantiated region summary in `StaticBranchChoiceGenerator.getNumSatChoices(DynamicRegion)`.