Adding heuristic to not instantiate regions when the instantiation does not reduce the region's branching factor
If the number of paths through the region is less than or equal to the number of choices that Java Ranger will explore through the region, then instantiating this region will not be beneficial, and so, this instantiation should be aborted. This change adds an estimator for the number of paths in `RegionMetricsVisitor` and gets the number of Java Ranger choices through the region in `StaticBranchChoiceGenerator.getNumSatChoices(DynamicRegion)`.

In this commit, I also make a small change to keep track of only the last exact count of paths in a region when performanceMode is on (in HeuristicManager).